lint fixing

+++ modules/core/include/core/util.hpp
@@ -43,8 +43,6 @@
 #ifndef _SSIG_CORE_UTIL_HPP_
 #define _SSIG_CORE_UTIL_HPP_
 
-#include <opencv2/core.hpp>
-
 #include <algorithm>
 #include <functional>
 #include <cctype>
@@ -52,6 +50,8 @@
 #include <string>
 #include <unordered_map>
 #include <vector>
+
+#include <opencv2/core.hpp>
 
 #include "core_defs.hpp"
 
@@ -136,7 +136,7 @@ class Util {
 
     fn["values"] >> values;
 
-    for (int i = 0; i < keys.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(keys.size()); ++i) {
       map[keys[i]] = values[i];
     }
   }
